feat: add and update transaksi

+++ app/src/main/java/com/stiproject/kelassti/model/response/transaksi/TransaksiData.kt
@@ -4,7 +4,7 @@ data class TransaksiData(
     val id: Int,
     val NIM_mahasiswa: Int,
     val nama: String,
-    val nominal: Int,
+    val nominal: Long,
     val type: String,
     val tanggal: String,
     val deskripsi: String
